requirements.txt
pytest
selenium
requests
assertpy
pytest-html
psycopg2-binary
